Implement one-time calculation of which modules are currently core.
Module::CoreList->is_core($pkg) is fine when you only need to inquire about
the core status of one module.  But since it recalculates its internal
per-perl-version lookup table with each invocation, it's not suitable for
inquiring about the status of multiple packages -- such as all the packages on
CPAN.

Let's assume that we are only concerned with whether a particular module is
core with one perl version -- typically, the version with which we're running
this program.  Let's further assume that that version (specified in 5.nnnnnn
format) is a well-populated element in $Module::CoreList::version.  If so,
then we can quickly determine a module's 'is_core' status and reduce the
running time of this program by a factor of 60.