Make "top_5_downstream" data consistent between runs on same data.
Until now, the listing of the "top 5 downstream" CPAN distributions for a
given distribution have been sorted in descending numerical order with no
provision for determining "ties".  This has meant that if you were to run the
program multiple times over the same data in the Mongo DB, you would almost
certainly get differences in the output.  Example:

    < 8656,Compress-Raw-Bzip2,cpan-upstream,PMQS,"IO-Compress:8655 File-Unpack:0 Task-BeLike-MELO:0 Task-DualLived:0"
    ---
    > 8656,Compress-Raw-Bzip2,cpan-upstream,PMQS,"IO-Compress:8655 Task-BeLike-MELO:0 Task-DualLived:0 File-Unpack:0"

In the above, the 3 reverse dependencies of Compress-Raw-Bzip2 which each have
0 revdeps of their own are presented in a different order from one run to the
next.

Providing a second-order sort in the Schwartzian Transform stabilizes this.

    8656,Compress-Raw-Bzip2,cpan-upstream,PMQS,"IO-Compress:8655 File-Unpack:0 Task-BeLike-MELO:0 Task-DualLived:0"