Estimate 'quasi-popularity' of each distribution.
The quasi-popularity of each distribution is a Boolean value indicating
whether or not at least one person other than the distro's uploaders is found
in the list of maintainers for any of the distro's revdeps.

Get the CPANID of the uploader from the MongoDB entry for each distro.

Get the maintainers of the revdeps from already existing lookup table.

Compare.  Set quasi-popularity (%qp) to true value if at least one revdep has
a maintainer other than the uploader.

TODO: Make use of this data while ordering the CPAN river.