Make variable names more self-documenting.
Extend use of command-line options to handle the name of the MongoDB (defaults
to 'cpan') and the MongoDB collection (defaults to 'packages').

Command-line option 'hacking' can be used during development to avoid all the
MongoDB-related stuff and just see what's happening with the module metadata.
'verbose' is also provided in a few places.